+++ app/utils/CheckYourAnswersHelper.scala
@@ -26,22 +26,7 @@ import uk.gov.hmrc.viewmodels._
 
 class CheckYourAnswersHelper(userAnswers: UserAnswers) {
 
-  def sealsInformation(sealIndex: Index): Option[Row] = userAnswers.get(SealsInformationPage) map {
-    answer =>
-      Row(
-        key     = Key(msg"sealsInformation.checkYourAnswersLabel", classes = Seq("govuk-!-width-one-half")),
-        value   = Value(yesOrNo(answer)),
-        actions = List(
-          Action(
-            content            = msg"site.edit",
-            href               = routes.SealsInformationController.onPageLoad(lrn, sealIndex, CheckMode).url,
-            visuallyHiddenText = Some(msg"site.edit.hidden".withArgs(msg"sealsInformation.checkYourAnswersLabel"))
-          )
-        )
-      )
-  }
-
-  def addSecurityDetails: Option[Row] = userAnswers.get(AddSecurityDetailsPage) map {
+   def addSecurityDetails: Option[Row] = userAnswers.get(AddSecurityDetailsPage) map {
     answer =>
       Row(
         key = Key(msg"addSecurityDetails.checkYourAnswersLabel", classes = Seq("govuk-!-width-one-half")),
