+++ app/utils/GoodsSummaryCheckYourAnswersHelper.scala
@@ -19,7 +19,7 @@ package utils
 import java.time.format.DateTimeFormatter
 
 import controllers.goodsSummary.routes
-import models.{CheckMode, LocalReferenceNumber, UserAnswers}
+import models.{CheckMode, Index, LocalReferenceNumber, UserAnswers}
 import pages._
 import pages.{AddCustomsApprovedLocationPage, AuthorisedLocationCodePage, DeclarePackagesPage, TotalGrossMassPage, TotalPackagesPage}
 import utils.GoodsSummaryCheckYourAnswersHelper.dateFormatter
@@ -28,6 +28,21 @@ import uk.gov.hmrc.viewmodels.Text.Literal
 import uk.gov.hmrc.viewmodels._
 
 class GoodsSummaryCheckYourAnswersHelper(userAnswers: UserAnswers) {
+
+  def sealsInformation(sealIndex: Index): Option[Row] = userAnswers.get(SealsInformationPage) map {
+    answer =>
+      Row(
+        key     = Key(msg"sealsInformation.checkYourAnswersLabel", classes = Seq("govuk-!-width-one-half")),
+        value   = Value(yesOrNo(answer)),
+        actions = List(
+          Action(
+            content            = msg"site.edit",
+            href               = routes.SealsInformationController.onPageLoad(lrn, sealIndex, CheckMode).url,
+            visuallyHiddenText = Some(msg"site.edit.hidden".withArgs(msg"sealsInformation.checkYourAnswersLabel"))
+          )
+        )
+      )
+  }
 
   def controlResultDateLimit: Option[Row] = userAnswers.get(ControlResultDateLimitPage) map {
     answer =>
