+++ conf/app.routes
@@ -264,10 +264,10 @@ POST       /:lrn/goods-summary/control-result-date-limit             controllers
 GET        /:lrn/goods-summary/change-control-result-date-limit      controllers.goodsSummary.ControlResultDateLimitController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
 POST       /:lrn/goods-summary/change-control-result-date-limit      controllers.goodsSummary.ControlResultDateLimitController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
 
-GET        /:lrn/goods-summary/:sealIndex/seals-information                        controllers.SealsInformationController.onPageLoad(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = NormalMode)
-POST       /:lrn/goods-summary/:sealIndex/seals-information                        controllers.SealsInformationController.onSubmit(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = NormalMode)
-GET        /:lrn/goods-summary/:sealIndex/change-seals-information                  controllers.SealsInformationController.onPageLoad(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = CheckMode)
-POST       /:lrn/goods-summary/:sealIndex/change-seals-information                  controllers.SealsInformationController.onSubmit(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = CheckMode)
+GET        /:lrn/goods-summary/:sealIndex/seals-information                        controllers.goodsSummary.SealsInformationController.onPageLoad(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = NormalMode)
+POST       /:lrn/goods-summary/:sealIndex/seals-information                        controllers.goodsSummary.SealsInformationController.onSubmit(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = NormalMode)
+GET        /:lrn/goods-summary/:sealIndex/change-seals-information                  controllers.goodsSummary.SealsInformationController.onPageLoad(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = CheckMode)
+POST       /:lrn/goods-summary/:sealIndex/change-seals-information                  controllers.goodsSummary.SealsInformationController.onSubmit(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = CheckMode)
 
 GET        /:lrn/goods-summary/:sealIndex/confirm-remove-seal                          controllers.goodsSummary.ConfirmRemoveSealController.onPageLoad(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = NormalMode)
 POST       /:lrn/goods-summary/:sealIndex/confirm-remove-seal                          controllers.goodsSummary.ConfirmRemoveSealController.onSubmit(lrn: LocalReferenceNumber, sealIndex: Index, mode: Mode = NormalMode)
